fix(tracing): Move the gateway handler to core and stamp env/host on telemetry
Consolidates #1690 into this PR so one change covers the whole "a failure has
to name itself in the backend" story.

ModelGatewayErrorHandler moves from packages/api to packages/core and registers
in Runner._get_api_app, so every app built from that base is covered rather than
only the main API. Its behavioural tests move with it; the api scope keeps the
one test only it can prove — that ApiRunner overriding _get_api_app still ends
up with the registration.

resource/deployment (from #1690) stamps deployment.environment and host.name on
every pipeline. SigNoz keys its Environment and Hostname quick-filters on those
two resource attributes and nothing set them, so both columns were empty for
every span and log in every environment — not just latest. Values come from
aihub-playbook's env.j2 (bbvch-ai/aihub-playbook#22), which has to land first.

The ${VAR:-default} in compose is deliberate and documented: an empty value
makes the processor fail to build and the collector exit, and nothing
depends_on it, so an unset var would drop all telemetry instead of one
attribute.

Corrections to the ADR, which described the old code accurately but the
installed libraries wrongly:

- opentelemetry-instrumentation-asgi 0.63b0 does mark the span. It wraps the
  stack in its own ExceptionHandlerMiddleware, and trace.use_span defaults to
  record_exception/set_status_on_exception. What was missing was never the
  ERROR status, only a record naming the cause. _mark_span_failed is still
  needed, for the opposite reason: a handled exception never reaches that
  branch.
- packages/bot does not fail the same way. BaseChatBot wraps _respond in
  try/except and OpenaiCompletionHandler translates APIStatusError in-route, so
  the app-level handler never sees it. Recorded as a trade-off instead: the bot
  reads the flat body["message"] while LiteLLM nests it, so it falls back to the
  "Error code: N - {...}" wrapper this PR exists to remove.

Verified: all 10 generated collector configs pass `validate` against
otel/opentelemetry-collector-contrib:0.154.0; core 55 unit tests; api openai
suite unchanged at 12 pre-existing infra-dependent failures.
