+++ docs/arc42/decisions/2026_08_19_model_gateway_error_translation.md
@@ -24,9 +24,13 @@ Three separate gaps turned that into an hour of blind searching:
 1. **The caller was told nothing.** Every model call in this API goes through the OpenAI SDK, whose exceptions are not
    `HTTPException`. Unhandled, they left the app as Starlette's plain-text 500 — no JSON, no message. OpenWebUI parses
    `error.message` out of the body, found none, and fell back to printing the transport error.
-2. **The trace was not marked.** `opentelemetry-instrumentation-asgi` has no exception branch in
-   `OpenTelemetryMiddleware.__call__` — only `try`/`finally` for metrics. An exception escaping the app ends the server
-   span with status `UNSET` and no `http.status_code`, so the request never appeared in any error-filtered view.
+2. **The trace named no cause.** The span itself *was* marked: `opentelemetry-instrumentation-fastapi` wraps the
+   middleware stack in its own `ExceptionHandlerMiddleware`, and `trace.use_span` defaults to
+   `record_exception=True` / `set_status_on_exception=True`, so an exception escaping the app already ended the server
+   span `ERROR`. What the span carried was the exception *type* — `BadRequestError` on `POST
+   /openai/audio/transcriptions` — and never `Invalid model name passed in model=whisper-large-v3`, because the SDK's
+   message is the wrapper `Error code: 400 - {…}`. An error-filtered trace view showed the request and still did not say
+   what to fix.
 3. **The traceback was not exported.** The OTLP `LoggingHandler` is attached to the root logger, but gunicorn's
    `UvicornWorker` re-parents `uvicorn.error` with its own handlers and `propagate = False`. The one record carrying the
    traceback — `Exception in ASGI application` — therefore stopped at the container's stdout and never reached the
@@ -65,9 +69,15 @@ PR is titled "…and audio endpoint error handling": it converted the API's *own
 ## Decision
 
 **Model-gateway failures are translated once, at the HTTP boundary, by `ModelGatewayErrorHandler`**
-(`packages/api/swiss_ai_hub/api/exceptions/model_gateway_error_handler.py`), registered in `ApiRunner._get_api_app` for
+(`packages/core/swiss_ai_hub/core/exceptions/model_gateway_error_handler.py`), registered in `Runner._get_api_app` for
 `APIStatusError`, `APITimeoutError` and `APIConnectionError`. Starlette resolves handlers along the exception's MRO, so
 every concrete SDK error is covered by those three registrations.
+
+It sits in `Runner` rather than `ApiRunner` because the boundary rule belongs to every app built from that base —
+`packages/sysadmin-api` and `packages/bot` both do — not to one service's runner. What it does **not** do is cover the
+bot's chat path: `BaseChatBot` wraps `_respond` in `try/except Exception` and `OpenaiCompletionHandler.handle_exception`
+already translates `APIStatusError` itself, inside the route. An app-level handler never sees those, by design of that
+class. See the trade-off below.
 
 **1 — The status is reclassified by whose fault it is, not copied.**
 
@@ -81,7 +91,9 @@ every concrete SDK error is covered by those three registrations.
 the OpenAI-compatible clients this API emulates read only `error.message`. Dropping either hides the cause from one of
 them. The message is the unwrapped upstream `error.message`, not the SDK's `Error code: N - {…}` wrapper.
 
-**3 — Every such failure marks the current span `ERROR`.** This deliberately departs from the OTel HTTP convention,
+**3 — Every such failure marks the current span `ERROR`.** Handling the exception is what makes this necessary: it no
+longer propagates out of the middleware stack, so the instrumentation's exception branch never sees it and the span would
+otherwise end with the status of a normal response. Marking it also deliberately departs from the OTel HTTP convention,
 which leaves 4xx unset on server spans on the assumption that a 4xx is the client's fault. For a gateway that passes
 upstream statuses through, that assumption is false: the incident above arrived as a 400. The SDK ignores a `set_status`
 back to `UNSET`, so the instrumentation's own status setter running afterwards cannot undo this.
@@ -112,7 +124,8 @@ the queryable attributes.
 - The incident's cause now reaches the end user's screen: OpenWebUI displays
   `External: Invalid model name passed in model=whisper-large-v3` instead of `500 Server Error`.
 - One registration covers every endpoint that talks to a model — chat, embeddings, images, STT, TTS — including the
-  dynamically registered agent endpoints.
+  dynamically registered agent endpoints, and any app built from `Runner` whose route lets the exception reach the
+  boundary.
 - Failures are visible in three independent signals with different failure modes: the response, the span, and stdout.
   Losing telemetry no longer loses the evidence.
 - Logs and traces are already correlated: `LoggingInstrumentor` injects `trace_id`/`span_id` into every record, so a log
@@ -132,6 +145,17 @@ the queryable attributes.
 - **The convention now has an exception that has to be taught.** "No defensive try-catch" still holds in services, but
   the boundary rule lives only in `packages/api/CLAUDE.md` and here. A new upstream integration that raises its own SDK
   exception type will repeat the original bug until someone registers it.
+- **`Runner` now knows about the OpenAI SDK, and pays for it at import.** The base class for every HTTP service gained a
+  dependency on one specific upstream client. `openai` was not previously in `Runner`'s import chain and takes ~750ms to
+  import, so a service that never calls a model (`sysadmin-api`) now pays that on cold start. Accepted because `openai`
+  is already a direct dependency of `packages/core`, but a second gateway SDK would make this the wrong place.
+- **The bot now has two translations of the same error, and the worse one wins.**
+  `OpenaiCompletionHandler.handle_exception` reads `exception.body["message"]`; this handler reads the nested
+  `body["error"]["message"]` that LiteLLM actually sends. The flat read misses, falls back to `exception.message`, and
+  shows the caller exactly the `Error code: N - {…}` wrapper this ADR exists to remove — while `BaseChatBot`'s
+  `except Exception` guarantees the app-level handler never gets the chance to do better. Pre-existing, not introduced
+  here, and not fixed here either: converging them means deciding whether the bot's in-route handling should exist at
+  all, which is a change to `BaseChatBot`, not to this boundary.
 - **Streamed failures mid-stream are still opaque.** Only the stream *opening* moved into the endpoint's scope. A
   provider that fails after the first token still truncates the response, and no status code can be sent by then.
 - **Server-logger export widens log volume.** `uvicorn.error` tracebacks now reach the paid backend. That is the point,
