+++ packages/api/CLAUDE.md
@@ -44,7 +44,6 @@ packages/api/
 │   ├── pagination/             # PageNumber, PageSize types
 │   ├── persistance/            # Event persistence to MongoDB
 │   ├── audio/                  # Audio chunking service
-│   ├── exceptions/             # Model-gateway error → HTTP response translation
 │   └── testing/                # Auth bypass utilities for tests
 ├── playground/testing/          # Interactive test server + test suite
 │   ├── main.py                 # Dev server entry point
@@ -126,18 +125,19 @@ All infrastructure clients are initialized in `runners/lifetime/lifetime_manager
 Services keep failing fast — they raise `HTTPException` for their own rejections (403, 404) and let everything else
 propagate. What must not propagate out of the app is a model-gateway error: every LLM/STT/TTS/embedding call goes
 through the OpenAI SDK, whose exceptions are not `HTTPException`, so unhandled they became Starlette's plain-text 500 —
-no cause for the caller, and an unmarked span for the observability backend (the ASGI instrumentation has no exception
-branch, so it never sees them).
+no cause for the caller, and a span naming only the exception type, never the upstream message that says what to fix.
 
-`ModelGatewayErrorHandler` (`exceptions/model_gateway_error_handler.py`, registered once in `ApiRunner._get_api_app`)
-translates them: it unwraps the upstream `error.message`, passes through only the statuses a caller can act on
+`ModelGatewayErrorHandler` (`packages/core/swiss_ai_hub/core/exceptions/model_gateway_error_handler.py`, registered once
+in `Runner._get_api_app` so every service inherits it) translates them: it unwraps the upstream `error.message`, passes
+through only the statuses a caller can act on
 (400/413/422/429) while mapping this deployment's own faults (401/403/404, all 5xx) to 502 and upstream timeouts to
 504. The response body carries the message under **both** `detail` and `error.message` — platform clients read the
 former, the OpenAI-compatible clients this API emulates (OpenWebUI, OpenAI SDKs) read only the latter.
 
-Every one of these failures marks the current span `ERROR`, deliberately including the 4xx that OTel's server-span
-convention would leave `UNSET`: passing an upstream status through means a 4xx here can just as well be a model name
-this deployment got wrong. Log level is the narrower signal — `ERROR` with the traceback for what an operator must fix,
+Every one of these failures marks the current span `ERROR`. Handling the exception is what makes that necessary — it
+never propagates out to the instrumentation's exception branch — and it deliberately includes the 4xx that OTel's
+server-span convention would leave `UNSET`: passing an upstream status through means a 4xx here can just as well be a
+model name this deployment got wrong. Log level is the narrower signal — `ERROR` with the traceback for what an operator must fix,
 `WARNING` for the three statuses they cannot act on (413/422/429), which also keeps a rate-limit storm from flooding
 the log pipeline.
