+++ packages/core/swiss_ai_hub/core/exceptions/__init__.py
@@ -1,0 +1,26 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from swiss_ai_hub.core.exceptions.model_gateway_error_handler import ModelGatewayErrorHandler
+
+__all__ = [
+    "ModelGatewayErrorHandler",
+]
+
+_LAZY_IMPORTS: dict[str, str] = {
+    "ModelGatewayErrorHandler": "swiss_ai_hub.core.exceptions.model_gateway_error_handler",
+}
+
+
+def __getattr__(name: str) -> object:
+    if name in _LAZY_IMPORTS:
+        import importlib
+
+        module = importlib.import_module(_LAZY_IMPORTS[name])
+        value = getattr(module, name)
+        globals()[name] = value
+        return value
+    msg = f"module {__name__!r} has no attribute {name!r}"
+    raise AttributeError(msg)
